Spring configurations changed to AutoConfiguration
