+++ src/main/java/org/qubership/integration/platform/designtime/catalog/configuration/MicroserviceRestTemplateAutoConfiguration.java
@@ -17,10 +17,10 @@
 package org.qubership.integration.platform.designtime.catalog.configuration;
 
 import org.jetbrains.annotations.NotNull;
+import org.springframework.boot.autoconfigure.AutoConfiguration;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
 import org.springframework.http.client.BufferingClientHttpRequestFactory;
 import org.springframework.http.client.ClientHttpRequestFactory;
 import org.springframework.http.client.SimpleClientHttpRequestFactory;
@@ -29,8 +29,8 @@ import org.springframework.web.client.RestTemplate;
 import java.time.Duration;
 import java.util.function.Supplier;
 
-@Configuration
-public class MicroserviceRestTemplateConfiguration {
+@AutoConfiguration
+public class MicroserviceRestTemplateAutoConfiguration {
 
     @Bean("restTemplateMS")
     @ConditionalOnMissingBean
